Extract "Accounts" component
